requirements.txt
apache-airflow==3.1.0
fastapi==0.120.0
joblib==1.5.2
numpy==2.3.4
pandas==2.3.3
scikit-learn==1.7.2
scipy==1.16.2
sqlalchemy==2.0.44
uvicorn==0.38.0
statsforecast
lightgbm
pyarrow
fastparquet
